requirements.txt
numpy
torch
scipy
pandas
matplotlib
tensorrt; sys_platform == "linux"
jupyter
tqdm
scikit-learn
PyYAML
